+++ e2e/fixtures/manageCases/caseWorker/dummyPayment/dummyPaymentAwpSubmitContent.ts
