+++ e2e/journeys/manageCases/createCase/dummyCase/dummyC100.ts
@@ -2,7 +2,7 @@ import { Page } from "@playwright/test";
 import { DummyC100ChildDetails } from "./dummyC100ChildDetails";
 import { C100SubmitAndPay } from "../C100SubmitAndPay/C100SubmitAndPay";
 import { DummyCreateInitial } from "./dummyCreateInitial";
-import { DummyPaymentConfirmation } from "../../caseWorker/dummyPaymentConfirmation";
+import { DummyPaymentConfirmation } from "../../caseWorker/dummyPayment/dummyPaymentConfirmation";
 
 interface dummyC100Options {
   page: Page;
