+++ e2e/pages/manageCases/caseWorker/dummyPayment/dummyPaymentConfirmationPage.ts
@@ -1,8 +1,8 @@
-import { CommonPage } from "../commonPage";
+import { CommonPage } from "../../commonPage";
 import { Page } from "@playwright/test";
-import { Selectors } from "../../../common/selectors";
-import { DummyPaymentConfirmationContent } from "../../../fixtures/manageCases/caseWorker/dummyPaymentConfirmationContent";
-import { Helpers } from "../../../common/helpers";
+import { Selectors } from "../../../../common/selectors";
+import { DummyPaymentConfirmationContent } from "../../../../fixtures/manageCases/caseWorker/dummyPayment/dummyPaymentConfirmationContent";
+import { Helpers } from "../../../../common/helpers";
 
 export class DummyPaymentConfirmationPage extends CommonPage {
   public static async dummyPaymentConfirmationPage(page: Page): Promise<void> {
